Tried basic routing
installed nodemon as a dev dependancy and now using it to run the app. added a new script to use it. created a routes folder and added an index.js file in it created routes 1 for get and 2nd for post request at / home page route. Tried using body-parser to parse data coming in incoming post request but it was giving warning body-parser is deprecated. Read about alternative and found that now we can use express.json() to parse json data in body so using it now